+++ Source/Widgets/Viewer/viewerCesiumInspectorMixin.js
@@ -27,9 +27,22 @@ define([
     /*global console*/
 
     /**
-
+     * A mixin which adds the CesiumInspector widget to the Viewer widget.
+     * Rather than being called directly, this function is normally passed as
+     * a parameter to {@link Viewer#extend}, as shown in the example below.
+     * @exports viewerCesiumInspectorMixin
+     *
+     * @param {Viewer} viewer The viewer instance.
+     *
+     * @exception {DeveloperError} viewer is required.
+     *
+     * @example
+     * // Add basic drag and drop support and pop up an alert window on error.
+     * var viewer = new Cesium.Viewer('cesiumContainer');
+     * viewer.extend(Cesium.viewerCesiumInspectorMixin);
      */
-    var viewerInspectorMixin = function(viewer, options) {
+
+    var viewerCesiumInspectorMixin = function(viewer) {
         if (!defined(viewer)) {
             throw new DeveloperError('viewer is required.');
         }
@@ -37,7 +50,7 @@ define([
         var cesiumInspectorContainer = document.createElement('div');
         cesiumInspectorContainer.className = 'cesium-viewer-cesiumInspectorContainer';
         viewer.container.appendChild(cesiumInspectorContainer);
-        var cesiumInspector = new CesiumInspector(cesiumInspectorContainer, viewer.cesiumWidget.scene, viewer.cesiumWidget.canvas);
+        var cesiumInspector = new CesiumInspector(cesiumInspectorContainer, viewer.cesiumWidget.scene);
 
         defineProperties(viewer, {
             cesiumInspector : {
@@ -48,5 +61,5 @@ define([
         });
     };
 
-    return viewerInspectorMixin;
+    return viewerCesiumInspectorMixin;
 });